move bar formatting related code to a new bar mod
